refactor(ChessWebController): Room 존재 여부 확인 로직 이동
